FIX: Example errors

+++ examples/PPMIn_example/ppmin_example.pde
@@ -17,7 +17,7 @@
 #define CHANNELS 4
 
 uint16_t g_values[CHANNELS];                  // output buffer for PPMIn
-uint16_t g_workIn[PPMIN_WORK_SIZE(CHANNELS)]; // we need to have a work buffer for the PPMIn class
+uint8_t  g_workIn[PPMIN_WORK_SIZE(CHANNELS)]; // we need to have a work buffer for the PPMIn class
 
 rc::PPMIn g_PPMIn(g_values, g_workIn, CHANNELS);
